# docs/openapi.yaml
openapi: 3.0.3
info:
  title: My Social Networks API
  version: 1.0.0
  description: |
    API REST du nouveau service d'événements et de groupes de Facebook (TP Node.js / Express / MongoDB).

    **Authentification** : `POST /auth/register` ou `POST /auth/login` renvoient un token JWT (valable 24h)
    à envoyer dans le header `Authorization: Bearer <token>`.

    **Format des erreurs** : toutes les erreurs ont la forme `{ code, message, errors }`, où `errors`
    est la liste des problèmes à corriger (ex : erreurs de validation).

    **Listes** : les listes sont paginées avec `?page=1&limit=20` (100 maximum) et renvoient
    `{ data, pagination: { page, limit, total, pages } }`.

    **Visibilité** :
    - un groupe *secret* ou un événement *privé* répond `404` à ceux qui n'en font pas partie (il « n'existe pas » pour eux) ;
    - un groupe *privé* est visible (nom, description) mais ses membres, son fil et ses événements sont réservés aux membres (`403`).
  contact:
    name: Saad Lagzouli
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Développement
...
security:
  - bearerAuth: []
paths:
...
  /groups:
    post:
      operationId: createGroup
      tags: [Groups]
      summary: Créer un groupe
      description: Le créateur devient administrateur et membre. Un fil de discussion est créé automatiquement.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/GroupInput' }
      responses:
        '201':
          description: Groupe créé
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Group' }
        '400': { $ref: '#/components/responses/BadRequest' }
    get:
      operationId: getGroups
      tags: [Groups]
      summary: Lister les groupes
      description: Les groupes secrets n'apparaissent que pour leurs membres. Accessible sans compte (groupes publics et privés).
      security: [{}, { bearerAuth: [] }]
      parameters:
        - { name: search, in: query, description: Recherche dans le nom, schema: { type: string } }
        - { name: type, in: query, schema: { type: string, enum: [public, private, secret] } }
        - { name: mine, in: query, description: Uniquement mes groupes, schema: { type: boolean } }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Liste paginée (sans les listes de membres)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GroupList' }
  /groups/{id}:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    get:
      operationId: getGroupById
      tags: [Groups]
      summary: Détail d'un groupe
      description: Les membres d'un groupe privé ne sont renvoyés qu'à ses membres.
      security: [{}, { bearerAuth: [] }]
      responses:
        '200':
          description: Groupe
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Group' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      operationId: updateGroup
      tags: [Groups]
      summary: Modifier les paramètres du groupe (administrateur)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/GroupInput' }
      responses:
        '200':
          description: Groupe modifié
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Group' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      operationId: deleteGroup
      tags: [Groups]
      summary: Supprimer le groupe (administrateur)
      description: Supprime aussi son fil de discussion. Ses événements continuent d'exister sans groupe.
      responses:
        '200': { description: Groupe supprimé }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /groups/{id}/members:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    get:
      operationId: getMembers
      tags: [Groups]
      summary: Membres du groupe
      security: [{}, { bearerAuth: [] }]
      responses:
        '200':
          description: Membres avec leur rôle
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/GroupMember' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      operationId: addMembers
      tags: [Groups]
      summary: Ajouter des membres (administrateur)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/GroupMembersInput' }
      responses:
        '200':
          description: Groupe mis à jour
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Group' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /groups/{id}/join:
    post:
      operationId: joinGroup
      tags: [Groups]
      summary: Rejoindre un groupe public
      description: Un groupe privé ou secret se rejoint uniquement sur ajout d'un administrateur (`403`).
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200':
          description: Membre ajouté
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Group' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /groups/{id}/members/{userId}:
    parameters:
      - $ref: '#/components/parameters/Id'
      - $ref: '#/components/parameters/UserId'
    patch:
      operationId: updateMemberRole
      tags: [Groups]
      summary: Changer le rôle d'un membre (administrateur)
      description: Le groupe doit garder au moins un administrateur (`409`).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role: { type: string, enum: [member, admin] }
      responses:
        '200': { description: Rôle modifié }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
    delete:
      operationId: removeMember
      tags: [Groups]
      summary: Quitter le groupe ou retirer un membre
      description: Un membre peut se retirer lui-même ; un administrateur peut retirer n'importe qui. Le dernier administrateur ne peut pas partir (`409`).
      responses:
        '200': { description: Membre retiré }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /groups/{id}/events:
    get:
      operationId: getGroupEvents
      tags: [Groups]
      summary: Événements du groupe
      security: [{}, { bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/Id'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Liste paginée
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EventList' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /groups/{id}/thread:
    get:
      operationId: getGroupThread
      tags: [Groups, Threads]
      summary: Fil de discussion du groupe
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200':
          description: Fil de discussion
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Thread' }
        '403': { $ref: '#/components/responses/Forbidden' }

  # ------------------------------------------------------------------ Events
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string, example: 6702f1c2a1b2c3d4e5f60718 }
    UserId:
      name: userId
      in: path
      required: true
      schema: { type: string }
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
  responses:
    BadRequest:
      description: Données invalides
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { code: 400, message: Bad Request, errors: ['end_date doit être après start_date'] }
...
    Forbidden:
      description: Action non autorisée pour cet utilisateur
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { code: 403, message: Forbidden, errors: ["action réservée aux organisateurs de l'événement"] }
    NotFound:
      description: Ressource introuvable (ou invisible pour cet utilisateur)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { code: 404, message: Event Not Found }
    Conflict:
      description: Conflit avec l'état actuel (doublon, plus de place, règle métier)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { code: 409, message: Conflict, errors: ["une personne ne peut obtenir qu'un seul billet par événement"] }
  schemas:
    Error:
      type: object
      properties:
        code: { type: integer }
        message: { type: string }
        errors: { type: array, items: { type: string } }
    Pagination:
      type: object
      properties:
        page: { type: integer, example: 1 }
        limit: { type: integer, example: 20 }
        total: { type: integer, example: 42 }
        pages: { type: integer, example: 3 }
    ObjectId:
      type: string
      example: 6702f1c2a1b2c3d4e5f60718
    Timestamps:
      type: object
      properties:
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }

    # ---------- users
    UserPublic:
      type: object
      properties:
        _id: { $ref: '#/components/schemas/ObjectId' }
        firstname: { type: string, example: Saad }
        lastname: { type: string, example: Lagzouli }
        avatar: { type: string, format: uri, nullable: true }
        bio: { type: string, nullable: true }
...
    GroupInput:
      type: object
      required: [name]
      properties:
        name: { type: string, maxLength: 100, example: Club rando EFREI }
        description: { type: string, maxLength: 2000, nullable: true }
        icon: { type: string, format: uri, nullable: true }
        cover_photo: { type: string, format: uri, nullable: true }
        type: { type: string, enum: [public, private, secret], default: public }
        allow_member_posts: { type: boolean, default: true, description: Les membres peuvent publier dans le fil }
        allow_member_events: { type: boolean, default: false, description: Les membres peuvent créer des événements dans le groupe }
    Group:
      allOf:
        - $ref: '#/components/schemas/GroupInput'
        - $ref: '#/components/schemas/Timestamps'
        - type: object
          properties:
            _id: { $ref: '#/components/schemas/ObjectId' }
            administrators: { type: array, items: { $ref: '#/components/schemas/UserPublic' }, description: Absent si le contenu est réservé aux membres }
            members: { type: array, items: { $ref: '#/components/schemas/UserPublic' }, description: Absent si le contenu est réservé aux membres }
            members_count: { type: integer }
            my_role: { type: string, enum: [admin, member], nullable: true }
            created_by: { $ref: '#/components/schemas/ObjectId' }
    GroupList:
      type: object
      properties:
        data: { type: array, items: { $ref: '#/components/schemas/GroupInput' } }
        pagination: { $ref: '#/components/schemas/Pagination' }
    GroupMember:
      allOf:
        - $ref: '#/components/schemas/UserPublic'
        - type: object
          properties:
            role: { type: string, enum: [admin, member] }
    GroupMembersInput:
      type: object
      required: [user_ids]
      properties:
        user_ids: { type: array, items: { $ref: '#/components/schemas/ObjectId' }, minItems: 1 }
        role: { type: string, enum: [member, admin], default: member }

    # ---------- events
    Settings:
      type: object
      properties:
        ticketing: { type: boolean, default: false, description: Billetterie (événement public uniquement) }
        shopping_list: { type: boolean, default: false }
        carpooling: { type: boolean, default: false }
    EventUpdate:
      type: object
      properties:
        name: { type: string, maxLength: 150, example: Soirée d'intégration }
        description: { type: string, maxLength: 5000, nullable: true }
        start_date: { type: string, format: date-time, example: '2026-11-14T19:00:00Z' }
        end_date: { type: string, format: date-time, description: Après start_date, example: '2026-11-15T02:00:00Z' }
        location: { type: string, maxLength: 300, example: 'EFREI, 30-32 avenue de la République, Villejuif' }
        cover_photo: { type: string, format: uri, nullable: true }
        visibility: { type: string, enum: [public, private], default: public }
        settings: { $ref: '#/components/schemas/Settings' }
...
    EventList:
      type: object
      properties:
        data: { type: array, items: { $ref: '#/components/schemas/EventUpdate' } }
        pagination: { $ref: '#/components/schemas/Pagination' }
...
    Thread:
      type: object
      description: Lié à un groupe OU à un événement, jamais aux deux
      properties:
        _id: { $ref: '#/components/schemas/ObjectId' }
        group: { type: object, nullable: true, description: Groupe du fil (ou null) }
        event: { type: object, nullable: true, description: Événement du fil (ou null) }
        messages_count: { type: integer }
        created_at: { type: string, format: date-time }
...